Changed type of ID for card (String) with java generics mechanism.

+++ org.tuleap.mylyn.task.core/src/org/tuleap/mylyn/task/internal/core/client/soap/TuleapSoapConnector.java
@@ -1063,7 +1063,7 @@ public class TuleapSoapConnector {
 			newComment = TuleapMylynTasksMessages.getString(TuleapMylynTasksMessagesKeys.defaultComment);
 		}
 		this.getTuleapTrackerV5APIPortType().updateArtifact(sessionHash, groupId,
-				artifact.getTracker().getId(), artifact.getId(),
+				artifact.getTracker().getId(), artifact.getId().intValue(),
 				valuesList.toArray(new ArtifactFieldValue[valuesList.size()]), newComment, UTF8);
 
 		monitor.worked(fifty);
